Code review updates
