+++ .rustfmt.toml
@@ -1,0 +1,12 @@
+group_imports = "StdExternalCrate"
+imports_granularity = "Module"
+newline_style = "Unix"
+normalize_doc_attributes = true
+overflow_delimited_expr = true
+reorder_impl_items = true
+use_field_init_shorthand = true
+wrap_comments = true
+format_code_in_doc_comments = true
+format_macro_matchers = true
+format_strings = true
+condense_wildcard_suffixes = true
