+++ pro-serde-versioned-derive/Cargo.toml
@@ -1,12 +1,10 @@
 [package]
-name = "versioning-derive"
+name = "pro-serde-versioned-derive"
 version = "1.0.0"
 edition = "2021"
 
 [lib]
 proc-macro = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 syn = { version = "2.0.15", features = ["extra-traits"] }
